+++ toxcore/tox.h
@@ -795,6 +795,7 @@ typedef enum {
     TOX_GR_INVALID
 } TOX_GROUP_ROLE;
 
+/* paralel to TOX_USERSTATUS */
 typedef enum {
     TOX_GS_NONE,
     TOX_GS_AWAY,
@@ -1019,7 +1020,7 @@ int tox_group_get_invite_key(const Tox *tox, int groupnumber, uint8_t *dest);
  * returns number of peers on success.
  * return -1 on failure.
  */
-int tox_group_get_names(const Tox *tox, int groupnumber, uint8_t nicks[][TOX_MAX_NAME_LENGTH], uint32_t lengths[],
+int tox_group_get_names(const Tox *tox, int groupnumber, uint8_t nicks[][TOX_MAX_NAME_LENGTH], uint16_t lengths[],
                         uint32_t num_peers);
 
 /* Returns the number of peers in groupnumber on success.
